refactor: Extracted contracts
