+++ factory/src/lib.rs
@@ -1,17 +1,32 @@
 use std::convert::TryFrom;
 use std::str::FromStr;
 
-use near_sdk::borsh::{
+use mintbase_deps::common::{
+    NFTContractMetadata,
+    StoreInitArgs,
+};
+use mintbase_deps::consts::{
+    GAS_CREATE_STORE,
+    GAS_ON_CREATE_CALLBACK,
+    NO_DEPOSIT,
+    STORE_STORAGE,
+};
+use mintbase_deps::interfaces::factory_self;
+use mintbase_deps::logging::{
+    NearJsonEvent,
+    NftStoreCreateLog,
+};
+use mintbase_deps::near_sdk::borsh::{
     self,
     BorshDeserialize,
     BorshSerialize,
 };
-use near_sdk::collections::LookupSet;
-use near_sdk::json_types::U128;
-use near_sdk::{
+use mintbase_deps::near_sdk::collections::LookupSet;
+use mintbase_deps::near_sdk::json_types::U128;
+use mintbase_deps::near_sdk::{
+    self,
     assert_one_yocto,
     env,
-    ext_contract,
     is_promise_success,
     near_bindgen,
     AccountId,
@@ -19,21 +34,7 @@ use near_sdk::{
     Promise,
     PublicKey,
 };
-
-use crate::common::{
-    NFTContractMetadata,
-    StoreInitArgs,
-};
-use crate::consts::{
-    GAS_CREATE_STORE,
-    GAS_ON_CREATE_CALLBACK,
-    NO_DEPOSIT,
-    STORE_STORAGE,
-};
-use crate::logging::{
-    NearJsonEvent,
-    NftStoreCreateLog,
-};
+use mintbase_deps::serde_json;
 // ------------------------------- constants -------------------------------- //
 
 // ----------------------------- smart contract ----------------------------- //
@@ -58,18 +59,6 @@ pub struct MintbaseStoreFactory {
 }
 
 // ----------------------- contract interface modules ----------------------- //
-#[ext_contract(factory_self)]
-pub trait OnCreateCallback {
-    fn on_create(
-        &mut self,
-        store_creator_id: AccountId,
-        metadata: NFTContractMetadata,
-        owner_id: AccountId,
-        store_account_id: AccountId,
-        attached_deposit: U128,
-    );
-}
-
 impl Default for MintbaseStoreFactory {
     fn default() -> Self {
         env::panic_str("Not initialized yet.");
@@ -283,7 +272,7 @@ impl MintbaseStoreFactory {
             .create_account()
             .transfer(self.store_cost)
             .add_full_access_key(self.admin_public_key.clone())
-            .deploy_contract(include_bytes!("../../../wasm/store.wasm").to_vec())
+            .deploy_contract(include_bytes!("../../wasm/store.wasm").to_vec())
             .function_call("new".to_string(), init_args, 0, GAS_CREATE_STORE)
             .then(factory_self::on_create(
                 env::predecessor_account_id(),
