+++ helper/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "helper"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+mintbase-deps = { path = "../mintbase-deps", features = ["helper-wasm"] }
