+++ mintbase-deps/src/common/payouts/splits.rs
@@ -5,7 +5,10 @@ use near_sdk::borsh::{
     BorshDeserialize,
     BorshSerialize,
 };
-use near_sdk::AccountId;
+use near_sdk::{
+    env,
+    AccountId,
+};
 use serde::{
     Deserialize,
     Serialize,
@@ -30,4 +33,24 @@ pub struct SplitOwners {
 // TODO: why defined here and then implemented inside store?
 pub trait NewSplitOwner {
     fn new(arg: SplitBetweenUnparsed) -> Self;
+}
+
+impl NewSplitOwner for SplitOwners {
+    fn new(split_between: HashMap<near_sdk::AccountId, u32>) -> Self {
+        assert!(split_between.len() >= 2);
+        // validate args
+        let mut sum: u32 = 0;
+        let split_between: HashMap<AccountId, SafeFraction> = split_between
+            .into_iter()
+            .map(|(addr, numerator)| {
+                assert!(env::is_valid_account_id(addr.as_bytes()));
+                let sf = SafeFraction::new(numerator);
+                sum += sf.numerator;
+                (addr, sf)
+            })
+            .collect();
+        assert!(sum == 10_000, "sum not 10_000: {}", sum);
+
+        Self { split_between }
+    }
 }
