--- mintbase-deps/src/indexer/mod.rs
@@ -1,29 +1,0 @@
-// // previously `mintbase_std`
-// pub use {
-//     bigdecimal,
-//     chrono,
-//     dotenv,
-//     futures,
-//     hyper,
-//     near_account_id,
-//     near_client_primitives,
-//     near_crypto,
-//     near_indexer,
-//     near_jsonrpc_client,
-//     near_jsonrpc_primitives,
-//     near_primitives,
-//     near_sdk,
-//     openssl_probe,
-//     rayon,
-//     reqwest,
-//     tokio,
-//     tokio_postgres,
-//     tokio_stream,
-//     tower,
-//     tracing,
-//     tracing_appender,
-//     tracing_subscriber,
-//     uuid,
-// };
-
-// ---------------------------- helper functions ---------------------------- //
