+++ test.sh
@@ -24,14 +24,16 @@ kill_the_damn_sandbox() {
 cargo +nightly fmt || fail "Formatting"
 cargo lint || fail "Linting"
 
-cargo check -p mintbase-deps --features store-wasm --message-format short || fail "Checking store"
-build_wasm store # needs to build here so that checking factory won't fail
+# prevent factory checking from failing
+touch wasm/store.wasm
 
+cargo check -p mintbase-deps --features store-wasm --message-format short || fail "Checking store"
 cargo check -p mintbase-deps --features factory-wasm --message-format short || fail "Checking factory"
 cargo check -p mintbase-deps --features helper-wasm --message-format short || fail "Checking helper"
 cargo check -p simple-market-contract --message-format short || fail "Checking market"
 cargo check -p mintbase-near-indexer || fail "Checking indexer"
 
+build_wasm store
 build_wasm factory
 build_wasm helper
 build_wasm market
